Doo update

+++ src/Samples/PhotoSky/PhotoSky/PSSkyDriveFolderPicker.m
@@ -2,7 +2,19 @@
 //  PSSkyDriveFolderPicker.m
 //  PhotoSky
 //
-//  Copyright (c) 2012 Microsoft. All rights reserved.
+//  Copyright 2014 Microsoft Corporation
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
 //
 
 #import "PSSkyDriveFolderPicker.h"
@@ -149,7 +161,7 @@
     
     // Configure the cell...
     PSSkyDriveObject *skydriveObj = [self.currentFolder.folders objectAtIndex:indexPath.row];
-	cell.textLabel.text = [NSString stringWithFormat:@"%@ (%d folders; %d files)", skydriveObj.name, skydriveObj.folders.count, skydriveObj.files.count];
+	cell.textLabel.text = [NSString stringWithFormat:@"%@ (%d folders; %d files)", skydriveObj.name, (int)skydriveObj.folders.count, (int)skydriveObj.files.count];
     
     return cell;
 }
